reActive/requirements.txt
botbuilder-core>=4.7.0
aiohttp
requests